Refactor dataset class selection and add num_workers argument
This commit refactors the dataset class selection in the `train_collie` function. If either `data_args.entity_type_masking_prob` or `data_args.negatives_prob` is set, the `CollieDatasetWithTransformations` class is used. Otherwise, the base `CollieDataset` class is used. Additionally, the `num_workers` argument is added to limit the number of workers to the minimum of the available CPU count and 16.